changed price to decimal field to allow for 2 digits decimal places

+++ backend/apps/v1/inventory/serializers/AbstractSerializers.py
@@ -5,7 +5,7 @@ class ItemSpecificationSerializer(serializers.Serializer):
     modelNumber = serializers.CharField()
     weight = serializers.FloatField()
     weightFormat = serializers.CharField()
-    price = serializers.FloatField()
+    price = serializers.DecimalField(max_digits=20,decimal_places=2)
     priceFormat = serializers.CharField()
     brandName = serializers.CharField()
